move git related functions into git_operation.py
